requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==2.10
iniconfig==2.0.0
kafka-python==2.0.2
multidict==6.0.4
packaging==23.1
pluggy==1.3.0
protobuf==4.24.3
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-mock==3.11.1
requests==2.25.1
urllib3==1.26.16
yarl==1.9.2
